+++ CHANGELOG.md
@@ -23,6 +23,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -   Inactive Wireless Transmitter model being emissive.
 -   Unneeded network graph updating after placing a network device.
+-   Cable blocks not updating connections properly when using wrench.
 
 ## [2.0.0-milestone.3.1] - 2023-10-30